Remove: user list api from activityFeedProvider without impacting current behaviour
