+++ openmetadata-ui/src/main/resources/ui/src/components/DataQuality/IncidentManager/TestCaseStatus/TestCaseIncidentManagerStatus.component.tsx
@@ -34,7 +34,6 @@ import { TestCaseStatusIncidentManagerProps } from './TestCaseIncidentManagerSta
 const TestCaseIncidentManagerStatus = ({
   data,
   onSubmit,
-  usersList,
   hasPermission,
 }: TestCaseStatusIncidentManagerProps) => {
   const [isEditStatus, setIsEditStatus] = useState<boolean>(false);
@@ -92,7 +91,6 @@ const TestCaseIncidentManagerStatus = ({
           data={data}
           open={isEditStatus}
           testCaseFqn={data.testCaseReference?.fullyQualifiedName ?? ''}
-          usersList={usersList}
           onCancel={onCancel}
           onSubmit={onSubmit}
         />
