+++ openmetadata-ui/src/main/resources/ui/src/components/DataQuality/IncidentManager/TestCaseStatus/TestCaseIncidentManagerStatus.interface.ts
@@ -11,12 +11,10 @@
  *  limitations under the License.
  */
 
-import { EntityReference } from '../../../../generated/entity/type';
 import { TestCaseResolutionStatus } from '../../../../generated/tests/testCaseResolutionStatus';
 
 export interface TestCaseStatusIncidentManagerProps {
   data: TestCaseResolutionStatus;
   onSubmit: (data: TestCaseResolutionStatus) => void;
-  usersList?: EntityReference[];
   hasPermission?: boolean;
 }
