+++ openmetadata-ui/src/main/resources/ui/src/components/DataQuality/TestCaseStatusModal/TestCaseStatusModal.interface.ts
@@ -10,7 +10,6 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-import { EntityReference } from '../../../generated/entity/type';
 import { TestCaseResolutionStatus } from '../../../generated/tests/testCaseResolutionStatus';
 
 export interface TestCaseStatusModalProps {
@@ -19,5 +18,4 @@ export interface TestCaseStatusModalProps {
   testCaseFqn: string;
   onCancel: () => void;
   onSubmit: (data: TestCaseResolutionStatus) => void;
-  usersList?: EntityReference[];
 }
