+++ openmetadata-ui/src/main/resources/ui/src/components/IncidentManager/IncidentManager.component.tsx
@@ -21,7 +21,7 @@ import React, { useCallback, useEffect, useMemo, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { Link, useHistory } from 'react-router-dom';
 import { WILD_CARD_CHAR } from '../../constants/char.constants';
-import { PAGE_SIZE_BASE, PAGE_SIZE_MEDIUM } from '../../constants/constants';
+import { PAGE_SIZE_BASE } from '../../constants/constants';
 import { PROFILER_FILTER_RANGE } from '../../constants/profiler.constant';
 import { usePermissionProvider } from '../../context/PermissionProvider/PermissionProvider';
 import { ResourceEntity } from '../../context/PermissionProvider/PermissionProvider.interface';
@@ -51,7 +51,6 @@ import {
 } from '../../rest/incidentManagerAPI';
 import { getUserAndTeamSearch } from '../../rest/miscAPI';
 import { searchQuery } from '../../rest/searchAPI';
-import { getUsers } from '../../rest/userAPI';
 import {
   getNameFromFQN,
   getPartialNameFromTableFQN,
@@ -63,10 +62,7 @@ import {
   getEpochMillisForPastDays,
   getStartOfDayInMillis,
 } from '../../utils/date-time/DateTimeUtils';
-import {
-  getEntityName,
-  getEntityReferenceListFromEntities,
-} from '../../utils/EntityUtils';
+import { getEntityName } from '../../utils/EntityUtils';
 import {
   getEntityDetailsPath,
   getIncidentManagerDetailPagePath,
@@ -134,9 +130,6 @@ const IncidentManager = ({
   const { getEntityPermissionByFqn, permissions } = usePermissionProvider();
   const { testCase: commonTestCasePermission } = permissions;
 
-  const [initialAssignees, setInitialAssignees] = useState<EntityReference[]>(
-    []
-  );
   const [isPermissionLoading, setIsPermissionLoading] = useState(true);
   const [testCasePermissions, setTestCasePermissions] = useState<
     TestCasePermission[]
@@ -361,28 +354,6 @@ const IncidentManager = ({
     }
   };
 
-  const fetchInitialAssign = useCallback(async () => {
-    try {
-      const { data } = await getUsers({
-        limit: PAGE_SIZE_MEDIUM,
-
-        isBot: false,
-      });
-      const filterData = getEntityReferenceListFromEntities(
-        data,
-        EntityType.USER
-      );
-      setInitialAssignees(filterData);
-    } catch (error) {
-      setInitialAssignees([]);
-    }
-  }, []);
-
-  useEffect(() => {
-    // fetch users once and store in state
-    fetchInitialAssign();
-  }, []);
-
   useEffect(() => {
     if (
       commonTestCasePermission?.ViewAll ||
@@ -493,7 +464,6 @@ const IncidentManager = ({
             <TestCaseIncidentManagerStatus
               data={record}
               hasPermission={hasPermission?.EditAll}
-              usersList={initialAssignees}
               onSubmit={handleStatusSubmit}
             />
           );
@@ -537,12 +507,7 @@ const IncidentManager = ({
         ),
       },
     ],
-    [
-      testCaseListData.data,
-      initialAssignees,
-      testCasePermissions,
-      isPermissionLoading,
-    ]
+    [testCaseListData.data, testCasePermissions, isPermissionLoading]
   );
 
   if (
